requirements.txt
certifi==2021.5.30
chardet==4.0.0
fake-useragent==0.1.11
idna==2.10
pymongo==3.11.4
pyTelegramBotAPI==3.7.9
python-dotenv==0.17.1
requests==2.25.1
schedule==1.1.0
urllib3==1.26.5
